Move MessageValue out of message/mod.rs
