+++ shotover-proxy/benches/benches/codec.rs
@@ -7,7 +7,8 @@ use criterion::{black_box, criterion_group, BatchSize, Criterion};
 use shotover_proxy::codec::cassandra::CassandraCodec;
 use shotover_proxy::frame::cassandra::{parse_statement_single, Tracing};
 use shotover_proxy::frame::{CassandraFrame, CassandraOperation, CassandraResult, Frame};
-use shotover_proxy::message::{IntSize, Message, MessageValue};
+use shotover_proxy::message::Message;
+use shotover_proxy::message_value::{IntSize, MessageValue};
 use tokio_util::codec::Encoder;
 
 fn criterion_benchmark(c: &mut Criterion) {
