+++ shotover-proxy/src/lib.rs
@@ -27,6 +27,7 @@ pub mod config;
 pub mod error;
 pub mod frame;
 pub mod message;
+pub mod message_value;
 mod observability;
 pub mod runner;
 mod server;
