+++ shotover-proxy/src/transforms/cassandra/sink_cluster/mod.rs
@@ -2,7 +2,8 @@ use self::node_pool::{NodePoolBuilder, PreparedMetadata};
 use crate::error::ChainResponse;
 use crate::frame::cassandra::{parse_statement_single, CassandraMetadata, Tracing};
 use crate::frame::{CassandraFrame, CassandraOperation, CassandraResult, Frame};
-use crate::message::{IntSize, Message, MessageValue, Messages, Metadata};
+use crate::message::{Message, Messages, Metadata};
+use crate::message_value::{IntSize, MessageValue};
 use crate::tls::{TlsConnector, TlsConnectorConfig};
 use crate::transforms::cassandra::connection::{CassandraConnection, Response};
 use crate::transforms::{Transform, TransformBuilder, Wrapper};
