+++ shotover-proxy/src/transforms/protect/crypto.rs
@@ -1,4 +1,4 @@
-use crate::message::MessageValue;
+use crate::message_value::MessageValue;
 use crate::transforms::protect::key_management::KeyManager;
 use anyhow::{anyhow, bail, Result};
 use chacha20poly1305::aead::rand_core::RngCore;
